Apply pre-transport review fixes + workspace lints + mise pin
Protocol tightening from the review:
- stream_epoch u8 -> u16 (eliminates the long-uptime collision risk)
- MonoNanos::now() switched from `as u64` to checked conversion
- Document why bincode (vs postcard / rkyv / prost) was chosen
- Document version-mismatch policy on ClientHello (fatal in v0)
- Document why Modifiers is sent on every key event (state-drift defence)
- Document aspect-ratio handling for normalised MousePosition
- TODO(security) note on decode() pending bincode Limit wiring

Pin `bincode = "=2.0.1"` exactly. 3.0 is out but unvetted; defer the
migration until we hit a need.

Workspace lints centralised: deny unused_must_use and
unsafe_op_in_unsafe_fn, warn rust_2018_idioms and the cast_* clippy
lints. Each crate opts in via `[lints] workspace = true`.

Toolchain pin via mise.toml at the repo root (`rust = "1.95.0"`,
matching the installed toolchain). MSRV in workspace.package stays at
1.80 - the floor we promise to keep compiling against.

All 7 protocol tests still green; clippy clean with -D warnings across
the workspace.

+++ crates/tether-protocol/src/video.rs
@@ -33,20 +33,24 @@ pub struct VideoFrameMeta {
 /// A single video datagram. Frames larger than the transport's max datagram
 /// size are sliced into multiple packets sharing `(stream_epoch, frame_seq)`.
 ///
-/// The `stream_epoch` field is bumped by the host whenever the encoder is
-/// restarted (resize, codec switch, HW context loss). Clients drop all
-/// packets from prior epochs.
+/// `stream_epoch` is `u16` (varint-encoded as 1 byte for typical values) so
+/// a long-lived session that restarts the encoder thousands of times can't
+/// collide with prior epochs.
+///
+/// The host bumps `stream_epoch` whenever the encoder is restarted
+/// (resize, codec switch, HW context loss). Clients drop all packets from
+/// prior epochs.
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub enum VideoPacket {
     First {
-        stream_epoch: u8,
+        stream_epoch: u16,
         frame_seq: u32,
         fragment_count: u16,
         meta: VideoFrameMeta,
         payload: Vec<u8>,
     },
     Continuation {
-        stream_epoch: u8,
+        stream_epoch: u16,
         frame_seq: u32,
         fragment_index: u16,
         payload: Vec<u8>,
